Switch the Prom-Fed and HPO charts to not publish.
After this change-set the PromFed chart will be built only via the PromFed repo.
This is mirroring the process we know and can easily manage in BRO.
Where the "upstream" chart/image produced by the repo match 1:1.

Then, when we implement branching it becomes easy to target the chart+image combo specific to the version of rancher we need in rancher/charts.